Bump iOS build to 27 and fix compilation and Apple Sign-In issues
- Increment build number to 27 for TestFlight release
- Fix DocumentPicker type mismatch (Long to UInt conversion)
- Enhance Apple Sign-In flow to manually trigger initialization after auth
- Add debug logging to Apple Sign-In authentication
- Increase Gradle heap to 8GB and disable Kotlin Native cache for iOS release builds
- Add export options plist for App Store distribution

+++ shared/src/iosMain/kotlin/com/together/newverse/util/DocumentPicker.kt
@@ -98,7 +98,7 @@ private class DocumentPickerDelegate(
             }
 
             // Convert to string
-            val cp850Encoding = CFStringConvertEncodingToNSStringEncoding(kCFStringEncodingDOSLatin1)
+            val cp850Encoding = CFStringConvertEncodingToNSStringEncoding(kCFStringEncodingDOSLatin1.toUInt())
             val content = NSString.create(data, cp850Encoding) as? String
             if (content == null) {
                 if (accessing) url.stopAccessingSecurityScopedResource()
